backend/pyproject.toml
[project]
name = "GameFleet"
version = "0.1.0"
description = "Test"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "docker>=7.1.0",
    "fastapi>=0.118.0",
    "mcstatus>=12.0.5",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.37.0",
]